requirements.txt
discord.py<1.5
pandas==1.0.4
matplotlib==3.2.1
